+++ test/test_components_artifactory.py
@@ -40,8 +40,7 @@ class TestArtifactory(unittest.TestCase):
         config = sb.objects.config.Config(version="1.0.0")
         args = argparse.Namespace(job="push", force=False, artifact='test', user='sean', token='abc123')
 
-        expectedException = """This artifact version has already been pushed.
-Please bump the version before pushing (skelebot bump) or force push (-f)."""
+        expectedException = "This artifact version already exists. Please bump the version or use the force parameter (-f) to overwrite the artifact."
 
         try:
             self.artifactory.execute(config, args)
@@ -88,12 +87,61 @@ Please bump the version before pushing (skelebot bump) or force push (-f)."""
         mock_input.return_value = "abc"
 
         config = sb.objects.config.Config(version="1.0.0")
-        args = argparse.Namespace(job="pull", version='0.1.0', artifact='test', user=None, token=None)
+        args = argparse.Namespace(job="pull", version='0.1.0', artifact='test', user=None, token=None, override=False)
 
         self.artifactory.execute(config, args)
 
         mock_artifactory.assert_called_with("artifactory.test.com/ml/test/test_v0.1.0.pkl", auth=("abc", "abc"))
         mock_open.assert_called_with("test_v0.1.0.pkl", "wb")
+
+    @mock.patch('skelebot.components.artifactory.input')
+    @mock.patch('builtins.open')
+    @mock.patch('artifactory.ArtifactoryPath')
+    def test_execute_pull_lcv(self, mock_artifactory, mock_open, mock_input):
+        mock_apath = mock_artifactory.return_value
+        mock_input.return_value = "abc"
+        mock_apath.glob.return_value = ["artifact_v1.1.0", "artifact_v0.2.4", "artifact_v1.0.0", "artifact_v2.0.1"]
+
+        config = sb.objects.config.Config(version="1.0.9")
+        args = argparse.Namespace(job="pull", version='LATEST', artifact='test', user=None, token=None, override=False)
+
+        self.artifactory.execute(config, args)
+
+        mock_artifactory.assert_called_with("artifactory.test.com/ml/test/test_v1.0.0.pkl", auth=("abc", "abc"))
+        mock_open.assert_called_with("test_v1.0.0.pkl", "wb")
+
+    @mock.patch('skelebot.components.artifactory.input')
+    @mock.patch('builtins.open')
+    @mock.patch('artifactory.ArtifactoryPath')
+    def test_execute_pull_lcv_not_found(self, mock_artifactory, mock_open, mock_input):
+        mock_apath = mock_artifactory.return_value
+        mock_input.return_value = "abc"
+        mock_apath.glob.return_value = ["artifact_v1.1.0", "artifact_v0.2.4", "artifact_v1.0.0", "artifact_v2.0.1"]
+
+        config = sb.objects.config.Config(version="3.0.9")
+        args = argparse.Namespace(job="pull", version='LATEST', artifact='test', user=None, token=None, override=False)
+
+        try:
+            self.artifactory.execute(config, args)
+            self.fail("Exception Not Thrown")
+        except RuntimeError as err:
+            self.assertEqual(str(err), "No Compatible Version Found")
+
+    @mock.patch('skelebot.components.artifactory.input')
+    @mock.patch('builtins.open')
+    @mock.patch('artifactory.ArtifactoryPath')
+    def test_execute_pull_override_and_lcv(self, mock_artifactory, mock_open, mock_input):
+        mock_apath = mock_artifactory.return_value
+        mock_input.return_value = "abc"
+        mock_apath.glob.return_value = ["artifact_v1.1.0", "artifact_v0.2.4", "artifact_v1.0.0", "artifact_v2.0.1"]
+
+        config = sb.objects.config.Config(version="0.6.9")
+        args = argparse.Namespace(job="pull", version='LATEST', artifact='test', user=None, token=None, override=True)
+
+        self.artifactory.execute(config, args)
+
+        mock_artifactory.assert_called_with("artifactory.test.com/ml/test/test_v0.2.4.pkl", auth=("abc", "abc"))
+        mock_open.assert_called_with("test.pkl", "wb")
 
     @mock.patch('skelebot.components.artifactory.input')
     @mock.patch('artifactory.ArtifactoryPath')
@@ -103,7 +151,7 @@ Please bump the version before pushing (skelebot bump) or force push (-f)."""
         path.exists.return_value = False
 
         config = sb.objects.config.Config(version="1.0.0")
-        args = argparse.Namespace(job="pull", version='0.1.0', artifact='test', user=None, token=None)
+        args = argparse.Namespace(job="pull", version='0.1.0', artifact='test', user=None, token=None, override=False)
 
         self.artifactory.execute(config, args)